+++ inst/include/cpp11eigen.hpp
@@ -4,8 +4,7 @@
 #include <omp.h>
 #endif
 
-#include <eigen/Dense>
-// #include <eigen/Eigenvalues>
 #include <cpp11.hpp>
+#include <eigen/Dense>
 #include <wrappers/matrices.hpp>
 #include <wrappers/vectors.hpp>