+++ NEWS.md
@@ -1,3 +1,8 @@
+# cpp11eigen 0.2
+
+* Wrappers for complex numbers.
+* Expanded range of tests.
+
 # cpp11eigen 0.1
 
 * First public version.